cronjob/requirements.txt
requests
jetpack-io
